useInput custom hook
