+++ src/pages/todoListsApp/atoms/inputWithSubmit/InputWithSubmit.js
@@ -4,13 +4,10 @@ import { useCallback } from "react";
 
 import s from "./inputWithSubmit.module.scss";
 import {
-  ADD_TODO_LIST,
-  ADD_TODO_ITEM,
-} from "../../../../store/todoListsApp/actions/actionTypes";
-import {
   addTodoList,
   addTodoItem,
 } from "../../../../store/todoListsApp/actions/todoActions";
+import useInput from "./hooks/useInput";
 
 const InputWithSubmit = ({
   addTodoList,
@@ -18,27 +15,11 @@ const InputWithSubmit = ({
   actionType,
   todoListId = null,
 }) => {
-  const [input, setInput] = useState("");
-
-  const changeHandler = useCallback((e) => setInput(e.target.value), []);
-  const submitHandler = useCallback(
-    (e) => {
-      if (input && (e.key === "Enter" || e.target.id === "addNew")) {
-        switch (actionType) {
-          case ADD_TODO_LIST:
-            addTodoList(input);
-            setInput("");
-            break;
-          case ADD_TODO_ITEM:
-            addTodoItem(input, todoListId);
-            setInput("");
-            break;
-          default:
-            break;
-        }
-      }
-    },
-    [actionType, addTodoItem, addTodoList, input, todoListId]
+  const [input, changeHandler, submitHandler] = useInput(
+    addTodoList,
+    addTodoItem,
+    actionType,
+    todoListId
   );
 
   return (
