+++ src/store/index.js
@@ -1,6 +1,8 @@
-import { createStore } from "redux";
+import { applyMiddleware, createStore } from "redux";
+// import thunk from "redux-thunk";
 import rootReducers from "./todoListsApp/reducers";
 
+// const store = createStore(rootReducers, applyMiddleware(thunk));
 const store = createStore(rootReducers);
 
 export default store;
